delete comment and this relationship
