+++ app/Http/Controllers/Admin/CommentController.php
@@ -5,26 +5,19 @@ namespace App\Http\Controllers\Admin;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Model\Comment;
-use DB;
 
 class CommentController extends Controller
 {
     /**
      * Delete a comment and relationship.
      *
-     * @param comment $id object comment
+     * @param Comment $comment object comment
      *
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Comment $comment)
     {
-        try {
-            Comment::findOrFail($id)->delete();
-            if (request()->ajax()) {
-                return response()->json(Response::HTTP_NO_CONTENT);
-            }
-        } catch (Exception $e) {
-            return response()->json(Response::HTTP_NOT_FOUND);
-        }
+        $result = $comment->delete();
+        return response()->json(400, 200);
     }
 }
