+++ public/js/dataTable-in-detail-post.js
@@ -1,5 +1,8 @@
+var commentTable;
+var id;
+
 $(function () {
-  $('#list-comments').DataTable({
+  commentTable = $('#list-comments').DataTable({
     'paging'      : true,
     'lengthChange': true,
     'searching'   : false,
@@ -7,4 +10,26 @@ $(function () {
     'info'        : true,
     'autoWidth'   : false
   })
+});
+
+$('.confirm-delete').on('click',function(e){
+    e.preventDefault();
+    id = $(this).data('id');
+})
+
+$('#delete-btn').click(function(){
+  token = $('meta[name="csrf_token"]').attr('content');
+  
+	$.ajax({
+    url : '/admin/comments/'+ id,
+		type : 'DELETE',
+		headers: { 'X-CSRF-TOKEN': token },
+		success: function(data){
+      commentTable.row('#comment-item-' + id).remove().draw(false);
+    },
+    error: function(error){
+      console.log('helo: '+error);
+    }
+	});
+	$('#confirm').modal('hide');
 })
